Restructure device tracker
